refactor: 移除手动builder代码并添加自动生成builder的指令
本次提交移除了多个域中的手动builder代码，并通过`go:generate`指令自动生成builder。这一改动提高了代码的可维护性，减少了重复代码，并确保builder的一致性。

+++ internal/application/service/judge/submit.go
@@ -49,7 +49,7 @@ func Submit(req request.JudgeReq, reqUser model.ReqUser) (int64, error) {
 		submission.WithUserId(int64(reqUser.Id)),
 		submission.WithProblemId(int64(req.ProblemId)),
 		submission.WithStatus(entity.JudgeIE), // 设置为IE状态,保证在错误导致中断的情况下显示为IE
-		submission.WithLength(uint32(len(req.SourceCode))),
+		submission.WithLength(int64(len(req.SourceCode))),
 		submission.WithLanguageId(int64(languageDomain.Id)),
 		submission.WithSourceCode(req.SourceCode),
 	)
@@ -126,7 +126,7 @@ func Submit(req request.JudgeReq, reqUser model.ReqUser) (int64, error) {
 	submissionDomain = submission.NewSubmission(
 		submission.WithId(submissionId),
 		submission.WithStatus(entity.JudgeStatus(runnerResult.Status.Id)),
-		submission.WithScore(score),
+		submission.WithScore(int64(score)),
 		submission.WithMemory(int64(runnerResult.Memory)),
 		submission.WithTime(runnerResult.Time),
 	)
